feat(verifiedEmail): move spam protection logic into verifiedEmail.services.ts
